# Decomposition Algorithm