@agnesLeroy 2022-02-21
Hi @orelcosseron! Before reviewing/merging this PR may we just check how it behaves in terms of performance? @aPere3 we have benches for the decomposition if I remember correctly, right?