@orelcosseron 2022-02-21
@agnesLeroy we discussed it with @aPere3: this new version adds operations and removes none, so there's no need for a bench, we already know the performance will be worse. But @aPere3 thinks the drop in performance will be minimal, and we're kinda ought to use this one.