@agnesLeroy 2022-02-21
@orelcosseron it would be good to have actual numbers I think (at least to show in a sprint demo). Maybe we could see the impact on the decomposition itself, then on the bootstrap? I think you can just use the benchmark crate for this, running the benches before and after your commit, criterion should tell you the performance differential.