@orelcosseron 2022-02-21
Do you want me to do that on AWS or any computer wiil do?