@agnesLeroy 2022-02-23
Benchmark results:

- the performance of the decomposition itself is worsened by 15%
- Regarding the bootstrap, the impact is minimal (the decomposition is not the most significant bottleneck in the PBS performance)

```
=========

Parameters:
-----------
        LweDimension(600),
        GlweDimension(1),
        PolynomialSize(1024),
        DecompositionBaseLog(7),
        DecompositionLevelCount(2),

Before:
-------
        32 time:   [12.375 ms 12.421 ms 12.453 ms]
        64 time:   [15.740 ms 15.794 ms 15.833 ms]

After:
------
        32 time:   [11.780 ms 11.861 ms 11.920 ms]
        64 time:   [15.398 ms 15.515 ms 15.654 ms]
```